Solved problem to reuse driver objects across multiple step definitions.
