+++ EntireCourse/src/test/java/Features/Login.feature
@@ -9,10 +9,6 @@ Scenario Outline: Validate Successful Login
 		|URL|
 		|http://www.facebook.com|
 		|http://www.gmail.com|
-@Test	
-Scenario: Validate Page load 
-	Given User is on another login page 
-		|http://www.youtube.com|
-	Then Home page is displayed 
+
