+++ EntireCourse/src/test/java/Features/Login2.feature
@@ -1,0 +1,6 @@
+Feature: Application_Login_2 
+@Test	
+Scenario: Validate Page load 
+	Given User is on another login page 
+		|http://www.youtube.com|
+	Then Home page is ok
